refactor: moved entity class file
